requirements.txt
alembic==1.6.2
aniso8601==9.0.1
attrs==21.2.0
blinker==1.4
boto3==1.17.69
botocore==1.20.69
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
limits==1.5.1
lorem-text==1.5
Mako==1.1.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
PyJWT==2.1.0
pyotp==2.6.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-slugify==5.0.2
pytz==2021.1
redis==3.5.3
s3transfer==0.4.2
sentry-sdk==0.20.3
six==1.16.0
SQLAlchemy==1.4.14
text-unidecode==1.3
timeago==1.0.15
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
